requirements.txt
marker-pdf
openai
numpy
pyvis
pandas
openpyxl
faiss-cpu
python-Levenshtein
markdown
weasyprint